backend/parser-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pytesseract==0.3.10
Pillow==10.2.0
pyzbar==0.1.9
beautifulsoup4==4.12.3
pypdf2==3.0.1
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
opencv-python==4.9.0.80
numpy==1.26.3
aiofiles==23.2.1
pytest==7.4.4
pytest-asyncio==0.21.1
httpx==0.25.2

# NLP & NER
spacy==3.7.2

